+++ agora/permissions.py
@@ -32,7 +32,7 @@ def get_rules():
         ('api/v2/services', 'retrieve', 'anonymous', '*', '*', '*', '*'),
         ('api/v2/services', 'create', 'superadmin', '*', '*', '*', '*'),
         ('api/v2/services', 'create', 'admin', '*', '*', '*', '*'),
-        ('api/v2/services', 'create', 'serviceadmin', '*', '*', '*', '*'),
+        ('api/v2/services', 'create', 'serviceadmin', '*', 'organisation_owned', '*', '*'),
         ('api/v2/services', 'update', 'superadmin', '*', '*', '*', '*'),
         ('api/v2/services', 'update', 'admin', '*', '*', '*', '*'),
         ('api/v2/services', 'update', 'serviceadmin', '*', 'owned', '*','*'),
@@ -329,6 +329,21 @@ def get_rules():
         ('api/v2/component-implementation-detail-links', 'partial_update', 'admin', '*', 'update_unique', '*', '*'),
         ('api/v2/component-implementation-detail-links', 'partial_update', 'serviceadmin', '*', 'update_owns_service_unique', '*', '*'),
         ('api/v2/component-implementation-detail-links', 'destroy', 'superadmin', '*', '*', '*', '*'),
-        ('api/v2/component-implementation-detail-links', 'delete', 'superadmin', '*', '*', '*', '*')
+        ('api/v2/component-implementation-detail-links', 'delete', 'superadmin', '*', '*', '*', '*'),
+
+        ('api/v2/organisations', 'list', 'superadmin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'list', 'admin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'list', 'serviceadmin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'list', 'observer', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'retrieve', 'superadmin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'retrieve', 'admin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'retrieve', 'serviceadmin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'retrieve', 'observer', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'create', 'superadmin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'update', 'superadmin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'partial_update', 'superadmin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'destroy', 'superadmin', '*', '*', '*', '*'),
+        ('api/v2/organisations', 'delete', 'superadmin', '*', '*', '*', '*')
+
     ]
     return rules