+++ tests/test_serviceadmin.py
@@ -49,6 +49,8 @@ def test_contact_information(serviceadmin, client, superadmin):
 def test_institutions(serviceadmin, client, superadmin):
     assertions_crud('institutions', serviceadmin, superadmin)
 
+def test_organisations(serviceadmin, client, superadmin):
+    assertions_crud('organisations', serviceadmin, superadmin)
 
 def test_components(serviceadmin, client, superadmin):
     """